+++ lib/api/unversioned/types.go
@@ -6,3 +6,13 @@ type TypeMetadata struct {
 	Kind       string `json:"kind" validate:"required"`
 	APIVersion string `json:"apiVersion" validate:"required"`
 }
+
+// ---- Metadata common to all resources ----
+type ObjectMetadata struct {
+	Name string `json:"name" validate:"name"`
+}
+
+// ---- Metadata common to all lists ----
+type ListMetadata struct {
+}
+
